# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Autenticação",
        "description": "Documentação da API de autenticação e API de gerenciamento de links",
        "contact": {
            "email": "gabriel.gomes@outlook.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/links": {
            "get": {
                "tags": [
                    "Links"
                ],
                "summary": "Listar Links",
                "description": "Retorna uma lista de links com opções de filtro.",
                "operationId": "761ce962469416df67c17e9be73e392a",
                "parameters": [
                    {
                        "name": "categoria_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "cidade_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bairro_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de links recuperada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "url": {
                                                        "type": "string",
                                                        "example": "https://example.com"
                                                    },
                                                    "descricao": {
                                                        "type": "string",
                                                        "example": "Descrição do link"
                                                    },
                                                    "categoria_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "cidade_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "bairro_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2024-10-03T15:51:05.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2024-10-03T15:51:05.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Links"
                ],
                "summary": "Cadastrar Link",
                "description": "Adiciona um novo link.",
                "operationId": "63b32ef30d8996005b2f8ca1e859ae96",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "url",
                                    "descricao",
                                    "categoria_id",
                                    "cidade_id",
                                    "bairro_id"
                                ],
                                "properties": {
                                    "url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://example.com"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Descrição do link"
                                    },
                                    "categoria_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "cidade_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "bairro_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Link cadastrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "url": {
                                                    "type": "string",
                                                    "example": "https://example.com"
                                                },
                                                "descricao": {
                                                    "type": "string",
                                                    "example": "Descrição do link"
                                                },
                                                "categoria_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "cidade_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "bairro_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-10-03T15:51:05.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-10-03T15:51:05.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erro de Validação"
                                        },
                                        "erros": {
                                            "properties": {
                                                "url": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "A URL é obrigatória"
                                                    }
                                                },
                                                "descricao": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "A descrição é obrigatória"
                                                    }
                                                },
                                                "categoria_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "A categoria é obrigatória"
                                                    }
                                                },
                                                "cidade_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "A cidade é obrigatória"
                                                    }
                                                },
                                                "bairro_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "O bairro é obrigatório"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Insira o token no formato (Bearer \"token\")",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}